Prepare for migration - move scala2 code which cannot be migrated to separate directory
