+++ core/src/main/scala-2/com/softwaremill/diffx/generic/auto/AutoDerivation.scala
