+++ core/src/main/scala-3/com/softwaremill/diffx/package.scala
@@ -1,0 +1,9 @@
+package com.softwaremill
+
+import scala.annotation.compileTimeOnly
+import scala.collection.Factory
+import com.softwaremill.diffx.DiffxSupport._
+
+package object diffx extends DiffxSupport {
+
+}
