+++ core/src/main/scalajvm/com/softwaremill/diffx/DiffxPlatformExtensions.scala
@@ -4,12 +4,12 @@ import java.time._
 import java.util.UUID
 
 trait DiffxPlatformExtensions {
-  implicit val diffUuid: Diff[UUID] = Diff.fallback
+  implicit val diffUuid: Diff[UUID] = Diff.useEquals
 
-  implicit val diffInstant: Diff[Instant] = Diff.fallback
-  implicit val diffLocalDate: Diff[LocalDate] = Diff.fallback
-  implicit val diffLocalTime: Diff[LocalTime] = Diff.fallback
-  implicit val diffLocalDateTime: Diff[LocalDateTime] = Diff.fallback
-  implicit val diffOffsetDateTime: Diff[OffsetDateTime] = Diff.fallback
-  implicit val diffZonedDateTime: Diff[ZonedDateTime] = Diff.fallback
+  implicit val diffInstant: Diff[Instant] = Diff.useEquals
+  implicit val diffLocalDate: Diff[LocalDate] = Diff.useEquals
+  implicit val diffLocalTime: Diff[LocalTime] = Diff.useEquals
+  implicit val diffLocalDateTime: Diff[LocalDateTime] = Diff.useEquals
+  implicit val diffOffsetDateTime: Diff[OffsetDateTime] = Diff.useEquals
+  implicit val diffZonedDateTime: Diff[ZonedDateTime] = Diff.useEquals
 }
