+++ specs2/src/test/scala/com/softwaremill/diffx/specs2/DiffMatcherTest.scala
@@ -1,7 +1,6 @@
 package com.softwaremill.diffx.specs2
 
 import org.specs2.Specification
-import com.softwaremill.diffx.generic.auto._
 
 class DiffMatcherTest extends Specification with DiffMatcher {
   override def is = s2"""This is an empty specification"""
