+++ utest/src/test/scala/com/softwaremill/diffx/utest/UtestAssertTest.scala
@@ -1,7 +1,6 @@
 package com.softwaremill.diffx.utest
 
 import utest._
-import com.softwaremill.diffx.generic.auto._
 
 object UtestAssertTest extends TestSuite with DiffxAssertions {
   val tests = Tests {
